fixed file not found bugs

+++ post-build.sh
@@ -16,21 +16,25 @@ if [ ! -d "common/utils" ]; then
     exit 1
 fi
 
-
-wget -O /tmp/human_model_files.tar \
+if [ ! -d "tools/" ]; then
+    echo "Downloading TOOLS ."
+    wget -O /tmp/human_model_files.tar \
     "https://huggingface.co/RendongZhang/GSAC-Dependencies/resolve/main/human_model_files.tar"
-tar -xf /tmp/human_model_files.tar -C common/utils
-rm /tmp/human_model_files.tar
+    tar -xf /tmp/human_model_files.tar -C common/utils
+    rm /tmp/human_model_files.tar
 
-wget -O /tmp/tools.tar \
-    "https://huggingface.co/RendongZhang/GSAC-Dependencies/resolve/main/tools.tar"
-tar -xf /tmp/tools.tar -C .
-rm /tmp/tools.tar
+    wget -O /tmp/tools.tar \
+        "https://huggingface.co/RendongZhang/GSAC-Dependencies/resolve/main/tools.tar"
+    tar -xf /tmp/tools.tar -C .
+    rm /tmp/tools.tar
 
-wget -O /tmp/sapiens.tar \
-    "https://huggingface.co/RendongZhang/GSAC-Dependencies/resolve/main/sapiens.tar"
-tar -xf /tmp/sapiens.tar -C tools/
-rm /tmp/sapiens.tar
+    wget -O /tmp/sapiens.tar \
+        "https://huggingface.co/RendongZhang/GSAC-Dependencies/resolve/main/sapiens.tar"
+    tar -xf /tmp/sapiens.tar -C tools/
+    rm /tmp/sapiens.tar
+   
+fi
+
 
 
 # Get the installation path of torchgeometry
@@ -50,3 +54,5 @@ if [ ! -d "$TARGET_DIR" ]; then
     echo "Error: Target directory $TARGET_DIR does not exist!"
     exit 1
 fi
+# Copy file1.py to the target directory
+cp -r conversions.py "$TARGET_DIR"
